added manage tasks

+++ migrations/20170828142413_add_track_table.js
@@ -1,16 +1,16 @@
 exports.up = function (knex, Promise) {
     return knex.schema.createTable('track', table => {
-        table.increments('id').primary();
-        table.integer('projectId').notNullable()
-        table.integer('taskId').notNullable()
+        table.bigIncrements('id').primary();
+        table.bigInteger('projectId').notNullable()
+        table.bigInteger('taskId').notNullable()
         table.string('date').notNullable()
         table.string('startTime').notNullable()
         table.string('endTime').notNullable()
         table.string('duration').notNullable()
         table.string('description').notNullable()
         table.string('additionalCost').notNullable()
-        table.integer('userID').notNullable()
-        table.integer('teamId').notNullable()
+        table.bigInteger('userID').notNullable()
+        table.bigInteger('teamId').notNullable()
         table.timestamp('createdAt').default(knex.fn.now())
         table.timestamp('timeStamp').default(knex.fn.now())  
     });
