+++ migrations/20170828142524_add_clients_table.js
@@ -1,13 +1,13 @@
 
 exports.up = function (knex, Promise) {
   return knex.schema.createTable('clients', table => {
-    table.increments('id').primary();
+    table.bigIncrements('id').primary();
     table.string('name').notNullable()
     table.string('email').notNullable()
     table.string('phone').notNullable()
     table.string('address').notNullable()
-    table.integer('userId').notNullable()
-    table.integer('teamId').notNullable()
+    table.bigInteger('userId').notNullable()
+    table.bigInteger('teamId').notNullable()
     table.timestamp('createdAt').default(knex.fn.now())
     table.timestamp('timeStamp').default(knex.fn.now())  
   });
